Revert "Homework (#17)"
This reverts commit 288377cddf184f4507598adc2ec6728807f54534.
