+++ src/com/company/homework2/serialization/figure/Shape.java
@@ -1,0 +1,5 @@
+package com.company.homework2.serialization.figure;
+
+public interface Shape {
+    void setArea();
+}
