--- src/main/java/com/holubinka/homework2/serialization/figure/Shape.java
@@ -1,5 +1,0 @@
-package com.holubinka.homework2.serialization.figure;
-
-public interface Shape {
-    void setArea();
-}
