--- src/main/java/com/holubinka/homework4/List.java
@@ -1,11 +1,0 @@
-package com.holubinka.homework4;
-
-public interface List<T> {
-    T get(int index);
-
-    void add(T t);
-
-    void remove(int index);
-
-    int size();
-}
